Create README.md
The project consists of developing and implementing a computer program that plays TIC TAC TOE against another player. This developed system will reduce the manual work that is for playing a game you’ve not to carry a pen or a pencil and a few sheets of papers on your pocket all the time. The system can able to provide hundreds or thousands of TIC TAC TOE game without any interruption.